Automatically provision personal namespace
Extend Workspace manager to support automatic provisioning
of personal namespaces.

The name of the namespace to create will be calculated from the
user name, which is expected to be an email address.

The user that namespace is created for will get konflux-admin
permissions on the namespace.

The implementation is idempotent.

This feature should be turned on using environment variables.

This change also adds an additional test suite. In order to run
it at the same time with the existing test suite (the default of
go test), allow to use different ports for workspace-manager.

In order to use the fakeclient provided by controller-runtime,
its version was updated.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>