Improved common test setup
1. Move common test setup to its own package so it can be shared
between test suites.

2. Improve the logic for starting the manager by:

- Separate the build and run commands. "go run" runs in its own process
so it's more difficult to find the process of "workspace-manger" and
kill it at the end of the test suite. This change fixing a bug
where the test suite didn't kill the workspace-manager process on exit.

- Wait for the server to be ready to serve http requests. For that
an additional endpoint /health was added to the server.

- Write the workspace-manager out to a log file.

3. Additional test suite was added for testing namespace provisioning.
It currently contains a dummy test. The actual test will be added
with the logic for provisioning namespaces. It was useful to add
and additional test suite (even a dummy one) in this change for verifying
the commons test setup code.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>